Tried to fix error msg displaying on local dev site. This concerns undefined variable in template html.tpl.php.
I linked to the web site from the template.php file. The site supplies the code used in the template.php file. There is a missing entry for the 'rdf_profile' variable so i have copied it in. If this works, will try to patch Jacine's github repository with the fix.